requirements.txt
bcrypt==3.1.7
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-admin-charts==0.11.0
django-admin-tools==0.8.1
django-autocomplete-light==3.5.1
django-bower==5.2.0
django-cache-utils==0.7.2
django-chart-tools==1.0
django-chartit==0.2.9
django-csvimport==2.13
-e git+https://github.com/mjs7231/django-dbbackup.git@d0e1f29e9d559a99753e8f219b6553b1345af381#egg=django_dbbackup
django-filter==2.2.0
django-hitcount==1.3.1
django-import-export==1.2.0
django-jsonfield==1.4.0
django-materializecss-form==1.1.17
django-nvd3==0.9.7
django-pandas==0.6.1
django-qsstats-magic==1.1.0
django-storages==1.8
django-taggit==1.1.0
django.js==0.8.1
djangorestframework==3.10.3
djcacheutils==3.0.0
et-xmlfile==1.0.1
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
joblib==0.14.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.0.3
mock==3.0.5
numpy==1.17.4
odfpy==1.4.0
openpyxl==2.6.4
pandas==0.24.2
paramiko==2.7.0
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pycparser==2.19
PyNaCl==1.3.0
pyparsing==2.4.5
python-dateutil==2.8.1
python-dotenv==0.10.3
python-memcached==1.59
python-nginx==1.5.3
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2019.3
PyYAML==5.1.2
scikit-learn==0.21.3
scipy==1.3.2
six==1.13.0
sklearn==0.0
SQLAlchemy==1.3.11
sqlparse==0.3.0
supervisor==4.1.0
tablib==0.14.0
Unidecode==1.1.1
Werkzeug==0.16.0
xlrd==1.2.0
xlwt==1.3.0
